Added a function to troubleshoot PruneLinksWorker and possible overlapping site reads

+++ app/workers/prune_links_worker.rb
@@ -16,25 +16,23 @@ class PruneLinksWorker < CoreWorker
     return false unless @domain = opts[:domain]
     @link_store = LinkQueue.new(domain: @domain)
     @temp_store = @link_store.members
-    @pruned_links = []
   end
 
   def perform(opts)
     return unless opts && init(opts)
     track
     while link = @temp_store.shift do
-      ld = LinkData.find(link)
+      unless ld = LinkData.find(link)
+        troubleshoot(ld)
+        next
+      end
       if !ld.listing_id && (listing = db { Listing.find_by_url(link) }) && listing.try(:fresh?)
-        @pruned_links << ld
+        ld.destroy
+        @link_store.rem(ld.url)
         record_incr(:links_pruned)
       else
         record_incr(:links_passed)
       end
-    end
-
-    @pruned_links.each do |ld|
-      ld.destroy
-      @link_store.rem(ld.url)
     end
 
     transition
@@ -46,5 +44,10 @@ class PruneLinksWorker < CoreWorker
     jid = ScrapePagesWorker.perform_async(domain: @domain)
     record_set(:transition, "ScrapePagesWorker")
     record_set(:next_jid, jid)
+  end
+
+  def troubleshoot(ld)
+    notify "TROUBLESHOOT: Missing linkdata: #{ld.to_h}"
+    notify "TROUBLESHOOT: Number of jobs for ScrapePagesWorker in flight with domain #{domain}: #{ScrapePagesWorker.jobs_in_flight_with_domain(domain).count}"
   end
 end
